@@ -10,22 +10,17 @@
     "json": "json-server --watch data/db.json --port 4000"
   },
   "dependencies": {
-    "@emotion/css": "^11.1.3",
-    "@emotion/react": "^11.4.0",
-    "@emotion/server": "^11.4.0",
-    "@emotion/styled": "^11.3.0",
-    "@material-ui/core": "^5.0.0-beta.2",
-    "@material-ui/icons": "^5.0.0-beta.1",
-    "@material-ui/lab": "^5.0.0-alpha.41",
-    "@material-ui/styles": "^4.11.4",
-    "@material-ui/system": "^5.0.0-beta.2",
+    "@emotion/react": "^11.11.3",
+    "@emotion/styled": "^11.11.0",
+    "@mui/icons-material": "^5.15.8",
+    "@mui/material": "^5.15.7",
     "axios": "^0.21.1",
     "formik": "^2.2.9",
     "framer-motion": "^4.1.17",
     "json-server": "^0.16.3",
-    "next": "11.0.1",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "next": "^14.1.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "yup": "^0.32.9"
   },
   "devDependencies": {
