@@ -1,4 +1,4 @@
-import { Container, Grid, Typography } from '@material-ui/core';
+import { Container, Grid, Typography } from '@mui/material';
 import MenuItemCard from '../components/MenuItemCard';
 import { menu } from '../data/menu-item';
 
